got it to compile, the only real problem was the SPI was being defined somewhere in the FT800 headers when it shouldn't have been. Sorted by making that definition extern and moving the real definition to screen.cpp
